# services/venue-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  venue-service_internal_dto.CreateSpaceRequest:
    properties:
      capacity:
        minimum: 1
        type: integer
      close_hour:
        type: string
      description:
        type: string
      name:
        type: string
      open_hour:
        type: string
      price:
        type: number
      type:
        enum:
        - private_office
        - meeting_room
        - desk
        type: string
    required:
    - capacity
    - close_hour
    - name
    - open_hour
    - price
    - type
    type: object
...
  venue-service_internal_dto.UpdateManagerRequest:
    properties:
      manager_id:
        type: integer
    required:
    - manager_id
    type: object
  venue-service_internal_dto.UpdateSpaceRequest:
    properties:
      capacity:
        minimum: 1
        type: integer
      close_hour:
        type: string
      description:
        type: string
      name:
        type: string
      open_hour:
        type: string
      price:
        type: number
      type:
        enum:
        - private_office
        - meeting_room
        - desk
        type: string
    type: object
...
host: localhost:8083
info:
  contact: {}
  description: This is the venue service API for the coworking booking system.
  title: Venue Service API
  version: "1.0"
paths:
...
  /spaces/{id}:
    delete:
      description: Delete a space by ID (user must be authenticated)
      parameters:
      - description: Space ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a space
      tags:
      - Space
    get:
      description: Retrieve detailed information of a space by ID
      parameters:
      - description: Space ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get space by ID
      tags:
      - Space
    put:
      consumes:
      - application/json
      description: Update information of a space (user must be authenticated)
      parameters:
      - description: Space ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateSpaceRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/venue-service_internal_dto.UpdateSpaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a space
      tags:
      - Space
  /spaces/{id}/manager:
    put:
      consumes:
      - application/json
      description: Assign or update manager for a space (user must be authenticated)
      parameters:
      - description: Space ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateManagerRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/venue-service_internal_dto.UpdateManagerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update manager of a space
      tags:
      - Space
  /spaces/search:
    get:
      description: Search spaces with filters and availability in a time range
      parameters:
      - description: Space name
        in: query
        name: name
        type: string
      - description: City
        in: query
        name: city
        type: string
      - description: Address
        in: query
        name: address
        type: string
      - description: Space type (private_office, meeting_room, desk)
        in: query
        name: type
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: start_time
        required: true
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: end_time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search spaces
      tags:
      - Space
...
  /venues/{id}/spaces:
    post:
      consumes:
      - application/json
      description: Create a new space under a specific venue (user must be authenticated)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      - description: CreateSpaceRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/venue-service_internal_dto.CreateSpaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new space under a venue
      tags:
      - Space
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"